#15 [MOD] DiagnosisLoading 수정
